+++ diesel_codegen/src/identifiable.rs
@@ -8,16 +8,18 @@ pub fn derive_identifiable(item: syn::MacroInput) -> Tokens {
     let table_name = model.table_name();
     let struct_ty = &model.ty;
     let lifetimes = model.generics.lifetimes;
-    let primary_key_name = model.primary_key_name;
+    let primary_key_names = model.primary_key_names;
     let fields = model.attrs;
-    if !fields.iter().any(|f| f.field_name.as_ref() == Some(&primary_key_name)) {
-        panic!("Could not find a field named `{}` on `{}`", primary_key_name, &model.name);
+    for pk in &primary_key_names {
+        if !fields.iter().any(|f| f.field_name.as_ref() == Some(pk)) {
+            panic!("Could not find a field named `{}` on `{}`", pk, &model.name);
+        }
     }
 
     quote!(impl_Identifiable! {
         (
             table_name = #table_name,
-            primary_key_name = #primary_key_name,
+            primary_key_names = (#(#primary_key_names),*),
             struct_ty = #struct_ty,
             lifetimes = (#(#lifetimes),*),
         ),