+++ diesel_codegen_syntex/src/model.rs
@@ -13,7 +13,7 @@ pub struct Model {
     pub attrs: Vec<Attr>,
     pub name: ast::Ident,
     pub generics: ast::Generics,
-    pub primary_key_name: ast::Ident,
+    pub primary_key_names: Vec<ast::Ident>,
     table_name_from_annotation: Option<ast::Ident>,
 }
 
@@ -26,9 +26,9 @@ impl Model {
         if let Annotatable::Item(ref item) = *annotatable {
             let table_name_from_annotation =
                 str_value_of_attr_with_name(cx, &item.attrs, "table_name");
-            let primary_key_name =
-                ident_value_of_attr_with_name(cx, &item.attrs, "primary_key")
-                    .unwrap_or(str_to_ident("id"));
+            let primary_key_names =
+                list_value_of_attr_with_name(cx, &item.attrs, "primary_key")
+                    .unwrap_or_else(|| vec![str_to_ident("id")]);
             Attr::from_item(cx, item).map(|(generics, attrs)| {
                 let ty = struct_ty(cx, span, item.ident, &generics);
                 Model {
@@ -36,7 +36,7 @@ impl Model {
                     attrs: attrs,
                     name: item.ident,
                     generics: generics,
-                    primary_key_name: primary_key_name,
+                    primary_key_names: primary_key_names,
                     table_name_from_annotation: table_name_from_annotation,
                 }
             })
@@ -45,8 +45,8 @@ impl Model {
         }
     }
 
-    pub fn primary_key_name(&self) -> ast::Ident {
-        self.primary_key_name
+    pub fn primary_key_names(&self) -> &[ast::Ident] {
+        &self.primary_key_names
     }
 
     pub fn table_name(&self) -> ast::Ident {